Fix bedrock2 wrapper functions
This is kind-of hacky

Note that we now also insert casts on the return value.  This probably
isn't required, but it seems better to do it, and I was previously not
doing it only because I didn't have easy access to the type.

+++ fiat-bedrock2/src/p224_32.c
@@ -2982,7 +2982,7 @@ uintptr_t internal_fiat_p224_nonzero(uintptr_t in0) {
 
 /* NOTE: The following wrapper function is not covered by Coq proofs */
 static void fiat_p224_nonzero(uint32_t* out1, const uint32_t arg1[7]) {
-  *out1 = internal_fiat_p224_nonzero((uintptr_t)arg1);
+  *out1 = (uint32_t)internal_fiat_p224_nonzero((uintptr_t)arg1);
 }
